Rename State switch resource string names
Previously the accessibility switch state announcement was saved under “on” as the resource name. This generic naming created conflicting resources (see b/302525864). The “on” and “off” resource strings have been renamed to “state_on” and “state_off,” which is more specific and will hopefully prevent further conflicting resources in the future.

Bug: 314195777

Test: NA, no logic change
Change-Id: I5b8d945502f731edb8dfc2ea2e8d80709fd86be2