Update ColorScheme to be Immutable
Based on the recommendation from API council, making the ColorScheme
immutable. This makes it more preformant on the general usage of the
scheme faster while making the individual color updates slower which is
uncommon.
Related CL: https://android-review.googlesource.com/c/platform/frameworks/support/+/2733678

Also added a ColorSchemeBenchmark to measure the overall change in
timings and allocations for such changes. Here are the numbers extracted
from running the benchmark with and without the change.

Before the change
  * Timing: ~2.8 ms (Avg of two runs)
  * Allocations: 1467 allocations
After the change
  * Timing: ~2.6ms (Avg of two runs)
  * Allocations: 1299 allocations
Based on the above results, we imporved nearly by 0.2 ms for the
initialization of the ColorScheme object and also reduced the overall
allocations that were being made for it.

Bug: 300090055
Relnote: """
ColorScheme is now Immutable, making individual color updates less efficient, but making more common usage of colors more efficient.

The reasoning behind this change is that the majority of apps wouldn't have updating individual colors as a main use case.

This is still possible but it will recompose more than before, in turn we significantly decrease the amount of state subscriptions through all of material code and will impact initialization and runtime cost of more standard use cases.
"""
Test: Existing tests

Change-Id: Ibc2d6d36f4bdd817c06f686cd1e3c210c2544b82