Convert ActivityResultRegistry to Kotlin.
Rewrite ActivityResultRegistry in Kotlin.

This ensures that the nullability in the
contract passed to register is correctly passed
through to the resulting nullability on the
ActivityResultLauncher returned.

Relnote: "`ActivityResultRegistry` has been rewritten
in Kotlin. This ensures that the nullability of the
generics on the contract passed to `register` will
correctly be passed through to the
`ActivityResultLauncher` returned to you."
Test: ./gradlew activity:activity:checkApi
BUG: 257291701

Change-Id: I121f0ca842066fb73ab0e51b89720029e991407a