Make CameraUseCaseAdapter only wraps one camera.
CameraUseCaseAdapter was designed to wrap multiple cameras and is able
to switch between them. However, the switch mechanism hasn't been
implemented and there's no case that will cause a CameraUseCaseAdapter
to switch to a different camera internally.

With a CameraConfig being introduced to CameraUseCaseAdapter, which may
internally alter the CameraControl and CameraInfo of the wrapped camera,
it's unclear how that should work with the multiple cameras that are
contained in the CameraUseCaseAdapter. We should probably redesign the
behavior if we want to support multiple cameras later on. So remove the
multiple camera field from CameraUseCaseAdapter for now.

Bug: b/310660777
Test: ./gradlew camera:camera-core:build
Change-Id: I3129704cc9cbcaf4cce35d98f256a252f634878a