Remove transform lastAccessedIndex after refresh
Transform lastAccessedIndex fulfills two roles - trigger prefetch and trigger load if new list is empty.

We still need to trigger new loads if refreshed page is empty, so we will still send an intial hint on empty refresh. However, we can rely on UI accessing items to update lastAccessedIndex. This delays prefetch by a frame but it is preferable over Refreshes requiring UI to return a transformed index.

Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew paging:paging-common:allTest
Bug: 315214786
Change-Id: I97ffed3e0f63a4551bc1196349157d2c30e8408b