Disable caching UnzipPaparazziNativeTransform
DisableCachingByDefault task is just an unzip, and that is a lot faster
to execute locally then fetch the results from the build cache. It is a
generally good practice in Gradle. E.g. Android Gradle Plugin marks
their ExtractAarTransform as @DisableCachingByDefault as well

Example of a slow fetch:
https://ge.androidx.dev/s/lkk6eqiqd6wmg/performance/transform-execution?details=fmrv5k4hq3ihq&expandedItem=WyIxMyJd&page=42
Remote download took 1m 47s, but actually running it is only ~1s

Test: None
Change-Id: I84ef882158d8a752b103681434cac781bed9fc86