Add retry to Window.captureToImage
Most of text field cursor test are experiencing flakiness due to PixelCopy request failure. According to the doc, we should retry if we get "no data" error.
This impl retries quite high in the code stack which ensures another redraw.

Also marked some of the ignored tests as flaky instead.

Test: added CaptureToImage test
Test: tests in CI should pass and ideally flaky tests shouldn't flake anymore
Bug: 301617059
Bug: 305799612

Change-Id: Ibe2f07d029dfda7f346444caf742ae932b045c9e