Fix memory leakage when Extensions are enabled.
When Extensions are enabled, if takePicture is not interrupted
before image is available, it is possible that the activity or
the fragment instance reference is held within OEM's implementation
due to CameraX's issue.

The reference chain is like
activity
1) --> indirectly held by SesscionConfig, DeferrableSurfaces
2) --> held by ProcessingCaptureSession
3) --> held by anonymous inner class for the captureCallback that was passed
    to OEM's implementation.

The problem is at 3) where anonymous inner class holds the
ProcessingCaptureSession instance. The solution is use a
static class instead.

Bug: 317754012
Relnote: "Fixed a memory leakage that results in activities
or fragments not being released when Extensions are enabled"
Test: manual

Change-Id: I142155b2b481007d5a45303e93992459b075d230