Add getSerializable compatible method
The `getSerializable()` method with no need of class parameter has been
deprecated since Android T. This CL fulfilled compatible methods into
`BundleCompat` and `IntentCompat`.

Note: The introduced method meets a requirement of type-safe API, it
returns the value if it exists and is of the `clazz` type, otherwise
returns null.

Bug: 317403466
Test: atest BundleCompat IntentCompat on T and U devices
Change-Id: Ib5565a69c83254a491e792117fc4984817df7444