Fix watch face library crash with GW prior to T KR2
GW with T was built against a library which was unconditionally
sending auxiliaryComponentPackageName leading to an
IllegalStateException in watch faces built with
wear-watchface-1.2.0. Newer WearServuces builds send null for
regular watch faces and the auxillary package name for resource
only watch faces (which is what the check was asserting), so it
makes sense for the the check to only apply from android U onwards.

Test: Manual testing
Bug: 318341363
Change-Id: I28e68e16c7a8be75652a8eaa93990d069a73c602