Implement PagingDataEvent equals, hashcode and string functions
Ideally we'd want PagingDataEvent to be a data class for how it is used. But data class is not recommended for public APIs because of backward-compat issues (https://kotlinlang.org/docs/jvm-api-guidelines-backward-compatibility.html#don-t-use-data-classes-in-an-api).

And data class cannot be sealed. So we just use a normal class with our own equals implementation.

Add tests to assert correct PagingDataEvent in PageStoreTest.

Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew paging:paging-common:allTest
Bug: 315214786
Change-Id: I5a6dee4f3c8182a7b513c91dde5a03331d199f6a