Update unused null check and add a test that GetSchemaResponse doesn't return empty sets.
Prior to ag/25766870, GetSchemaResponse would keep these entries and
they would appear in the public API. After that CL, such entries are
no longer retained.

The change in AppSearchImpl in this CL is not required because of
ag/25766870, but the != null check that is being replaced in this CL
could never trigger and it's nice to avoid calling unneeded APIs.

This also makes the behaviour consistent in case ag/25766870 needs to be
rolled back.

Test: AppSearchSessionCtsTestBase
Change-Id: I2cef2bb8b5f55a4ca5ec2f1dc2e8f0b6c92ccfef