View Sharing from Sdk in App for UiLib Test App
Adds the functionality in the jetpack Ui Lib Test App
to allow the mediator SDK in SdkRuntime to retrieve a
Bundle representing an adapter shared by Mediatee SDK
in the App process. This bundle will be used by
Mediator in the SDKRuntime to create a session that will
be shared with the app.

Test: manual test
Bug: 311164353
Change-Id: I3cc6b1ea301e82e886caccb144577d6de70686ba