Use compat API and delete API22 implementation
There's a handy API for setting an asynchronous bit on posted
messages. The platform API was introduced in API22, which is why
there is a JankStats implementation for that version. But there's
also a compat API in MessageCompat that handles the functionality
for earlier APIs as well, which is both more correct and cleaner
than JankStats' current implementation. Using this compat API
allows us to delete the API22 implementation class since this is
the only reason it was needed.

Bug: 247810198
Test: JankStats passes on releases before and after API22
Change-Id: Ic96ec809c51f7999deffc474eb2344526c3a96ec