Add AffiliationName to Jetpack CredEntry APIs
This adds on a new parameter that developers can use to affiliate
entries with a certain name. The extended API surface affects the
'CredentialEntry' abstract object, with a new default 'null' value. The
main subtype utilizing this new property is the
'PasswordCredentialEntry', which are the password entries that show up
on the CredentialManager Selector UI.

Bug: 319455779
Test: JUnit Tests Chained In
Relnote: "Adds a new affiliationName property to the CredentialEntry API
surface, and the PasswordEntry subclass surface."

Change-Id: I6261ec1d8cc36b2d96b284014d4ba4bc2773c1e3