Fix snapTo() target state changes
Fixes: 319714894
Fixes: 325454563

When SeekableTransitionState.snapTo() is followed immediately
by an animateTo(), the target value was not being changed before
animating to the new value.

This CL changes snapTo() to seekTo() and now waits for the
composition to finish before seekTo() completes. A new method
snapTo() is added to handle instantaneous changes to the target
state.

RelNote: "SeekableTransitionState has been updated
to enable transitioning between multiple states
instead of two states. Seeking can be done with
seekTo() and snapTo(). Animation can be
done with animateTo().

When a transition is interrupted, the initial
values will start or continue to animate to the
former targetState values."

Test: manual testing, new test
Change-Id: I0ef6b9d81af512fbf0fc95bb4672d607cc3b24e0