Add a workaround for a Skia issue on drawing arcs.
Here we add a small delta to arcs with length equal to 180 degrees. This covers all cases we have identified so far and but it's not an exhaustive suppression of the issue.

Since the "Tail" part of the arc line is always drawn in a lower layer, this extra segment is not visible. For the "Head" part, a small amount of length is added (which seems to be small enough not to be caught on screenshot tests).

Bug: 318337974
Relnote: Add a workaround for a skia arc drawing issue.
Change-Id: I08f09eb58257228bb74aba3319a8510fb85d57f8