Added LocalTextLinkStyle for links styling
This CLs adds composition local to style links similar to TextSelectionColors.
The value is read inside foundation (similarly to the selection colors) which means that even if calling BasicText wrapped by MaterialTheme, the links will be styled with Material specs.

See go/compose-hyperlinks-api for details and alternatives.

This CL doesn't set the corresponding values in material libraries, wear and TV. Those changes will be done separately.

Test: new tests + demo
Relnote: "Added LocalTextLinkStyle composition local that allows to change the style of the links in Text across the app. If you're using your own theme in the app, you should be setting this composition local according to your theming. When using Material theme the color of the link by default will be set to Material's primary color."

Change-Id: I7eb10b55fcec1b268c60744de700058eb11a385a