Revoke "Change Android V overlay document to take proto."
This reverts commit ece8c82cad8a622077a5543107859f6ca2b850ad.

Reason for revoke: Fix the setSchema fail by force override to true.

Original reason for revert: Changing the schema of the AndroidVOverlay document will cause setSchema to fail on any build created after Jan 12.
This needs to be updated to call setSchema with forceOverride=true if that specific situation is detected.

Change-Id: Iea76e78eec9b6db65e38fc7516a1081f33dbe1d6