Add new pill shapes (plain and star)
A pill shape has two semicircles at the elongated ends
with straight edges between them. A starPill has that same
countour, but with inner and outer radii and vertices, like
a star shape, but on the pill contour instead of a circular
contour.

Bug: 319723277
Test: Standard shapes unit and integration tests pass
Relnote: Added new pill and pillStar shapes
Change-Id: I71827916f76e5e2d10a0cce1def7440a2e2b8eca