Fix compatibility mode callbacks when new room-runtime is combined with old gen code.
Specifically, when old generated code is used with new room-runtime there is no OpenDelegate and instead the runtime is forced to use RoomOpenHelper which invokes the various RoomDatabase.Callback provided and stores a reference to the SupportSQLiteDatabase for isOpenInternal() and other APIs. When the connection manager was added, compatibility mode was also added but it missed extracting SupportSQLiteDatabase from the generated delegate used in the RoomOpenHelper which in turn broke invalidation tracker. This CL fixes the problem by installing a onOpen callback so that the connection manager can extract the database reference. It also moves callback invocation to the RoomOpenHelper instead of the generated code as RoomDatabase.mCallbacks is no longer set during initialization.

To further protect KMP changes the a flavor was added to Room's Java integration app so that all tests are run with a pre-KMP version of the annotation processor but with the new runtime. This scenario is likely to occur with an app using WorkManager and Room together.

Bug: 299168035
Bug: 316636123
Bug: 321297252
Test: ./gradlew :room:integration-tests:room-testapp:cAT
Change-Id: Ib139938e8f0e9557678432d36e7e284e8b257ad3