Narrow the scope of making selected nodes not clickable for a11y
Currently, any node that is selected is not conveyed as clickable to
a11y. This correctly applies to RadioButtons and tabs, where clicking on
an already selection radio button or tab will do nothing.

But chips, specifically FilterChips, can be clicked while selected (to
be unselected). To convey this functionality to accessibility services,
we only don't add a click action if the node is a tab or radiobutton

Relnote: Fixed an a11y bug allowing non-tabs and non-radiobuttons to be clickable when selected

Test: manual with Material compose demo app and FilterChips, tabs, and radio buttons, AndroidAccessibilityTest
Bug: 322242340
Change-Id: I2181c38693bd70e755a7514d655ef9091eef0975