Refactor defaults pattern for Card and ToggleButton
Based on the advice from the API council, this change refactors the
defaults pattern for Card and ToggleButton. The idea is to reduce the
use of CompositionLocals by removing the call to `.value` from the
default color methods.

To accomplish this, we are adding a new API which will reuse the cached
instances instead of calling the CompositionLocal for each invocation.

Bug: 315289418
Test: Existing tests
Relnote: """
Refactor defaults pattern for CardColors, ToggleButtonColors and
SplitToggleButtonColors by creating cached instances internally and
reducing the usage of CompositionLocal.
"""

Change-Id: If3fece36743a67fda21bb2c11533d70093568802