Update Lazy Scrollables Benchmark to include information about Scrollable layer.
Benchmark for LazyScrollables didn't include information about the entire scrolling cycle (start and finish were not being measured). In this CL we're updating these benchmarks to include measurements from the complete scrolling cycle. That way we can also use them to measure improvements in the Scrollable layer.

Relnote: N/A
Test: Benchmarks should run normally.
Change-Id: I0796eb9f12feeb8151282f0ffd340c6578c652cf