Fix API breakages around protected Subject constructors
Note to API council - this is an unpublished library, so please ignore
this for API review. API tracking is only enabled to check against
compatibility with the public API of Truth.

Seems most Subjects have protected constructors to allow subclassing,
but it was essential to add an internal overload in Kotlin since
protected in Kotlin does not allow same-module / package-private calls,
which we rely on.

NOTE - This breaks named arguments in internal source, but public usage
is fine.

Relnote: "N/A"
Test: ./gradlew kruth:kruth:checkApi
Change-Id: I39588d7d148592520427176f67fcd96485be795f