[Badge] Updated Alignment
Updated the badge alignment to be 6.dp closer to the center of the anchor in respect to the top-right corner of the anchor in the case when there is no content for the badge.
Updated the badge alignment to be 12.dp closer horizontally and 14.dp closer vertically to the center of the anchor in respect to the top-right corner of the anchor in the case when there is content for the badge.

Bug: 323532530
Test: Screenshot tests updated due to the updated alignment. Updated positional unit tests with the new alignment offsets.
Relnote: The badge alignment is adjusted to be closer to the center of the anchor content in respect to the top-right corner.
Change-Id: I18a28437cf249cd7d7ddffe103c89bdbe200399a