Disable charging experience for wear macrobenchmarks
This is an experiment to see if disabling the charging experience fixes
the flakiness of wear macro benchmarks. As of now, wear macrobenchmarks
are failing for r11 targets because the charging screen takes over the
display and components required for macro benchmark tests are no longer
present during the test.

Although it is difficult to see if making this change really helps with
the flakiness in the continuous integration, if this succeeds, the next
step would be to move this to a different common place. If this does not
work, we may need to go back and see what else needs to be done to fix
the flakiness.

Bug: 189952249
Test: Existing tests
Change-Id: I75559e1c615da94d02ab69c7cb0b96ab64e338cc