Update LazyListFocusMoveTest to use ParameterizedComposeTestRule
Update LazyListFocusMoveTest to use ParameterizedComposeTestRule. This change splits the parameters in this test into two sets, one of them with layout direction and reverse layout info can be run using the switch of composable parameters and ParameterizedComposeTestRule. The other set with just the focus move direction can continue to be parametereized. This reduced the number of tests to run from 256 to 64, but kept the same coverage. The total runtime of this test reduced by 58%.

Relnote:  N/A
Bug: N/A

Test: Previous tests should pass.
Change-Id: I35d1b64eac0fcbe48bb557892087e44deb04aeeb