Add shouldPublishDocs to build_info
Corresponding field was added to jetpad in cl/605602546.

This uses the same rules as the task for docs-tip-of-tree, a project publishes docs if it runs API tasks or is a samples project, and doesn't specify a `doNotDocumentReason`.

Because separate build_info files are generated for each platform of a KMP project but only the one without a platform suffix is listed in docs-public/build.gradle, this sets shouldPublishDocs to false for the platform-specific build_info files.

I ran `./gradlew createAggregateBuildInfoFiles` and checked some different cases were correct:
* androidx.activity_activity_build_info.txt - publishes docs
* androidx.appactions.interaction_interaction-proto_build_info.txt - uses `doNoDocumentReason`
* androidx.benchmark_benchmark-gradle-plugin_build_info.txt - project type does not run API tasks
* androidx.collection_collection_build_info.txt - non-platformed artifact for KMP project (the one used for docs)
* androidx.collection_collection-jvm_build_info.txt - platformed artifact for KMP project (not used for docs)
* androidx.activity_activity-compose-samples_build_info.txt - samples project

Test: CreateLibraryBuildInfoFileTaskTest, manual checks listed above
Bug: 289373390

Change-Id: Ie0582bbc8c534ed954ff1b883855653e5a6f91a6