Check the prevInflatedParent before centering
When doing a diff update, we center previously inflated parent if
exists, by changing the gravity in FrameLayout.LayoutParams.

This works correctly when that View wasn't changed from the
outside, as inside the renderer we are sure it will be class of
FrameLayout.LayoutParams. However, if this element was changed
from the outside before diff update to another type of View and
LayoutParams, this will cause ClassCastException.

To prevent that, we are checking if it's the parent we provided
before centering.

Bug: 309548879
Test: Added
Change-Id: I336186bc4b853cc4e05c59ab7822b9f58d2c8245