Implement TruthFailureSubject + ExpectFailure
TruthFailureSubject is the last class to migrate, which finally enables
migrating a large suite of ExpectFailure tests.

In order to get these tests passing, it was necessary to make a bunch of fixes
for string representation of chained subjects, especially allowing string
overrides and specific behaviors around printing names of classes.

Relnote: "N/A"
Test: ANDROIDX_PROJECTS=INFRAROGUE ./gradlew kruth:kruth:allTests -Pandroidx.enabled.kmp.target.platforms=+LINUX
Change-Id: I500ebb61bef4b8e4781d638d3fa505cf5170e59e