Enable `nonSkippingGroupOptimization` for compose libraries
This enables for the removal of the function group for
functions that are not skippable or restartable that was
added in aosp/2912628.

This change includes runtime fixes for the ordering of
`onForgotton`, `onRelease` and `onDeactivate` when the
compiler switch is enabled. The ordering of `onDispose`
calls for `DisposableEffect` for example, may be out of
order in runtimes prior to this fix  when the switch is
enabled.

Test: ./gradlew :compose:r:r:tDUT
Bug: 254480106
Change-Id: I486ef31e83f76763ffb3d3ca6ff78f04e7b0e465