Fix the race condition in SurfaceEdge.
There is a race condition in SurfaceEdge that is triggered when Compose is used. There are 2 actions that can be racy: 1) when the SurfaceEdge disconnects, it closes the current SettableSurface and creates a new one. 2) SurfaceEdge closes the SettableSurface if the downstream node is closed.

The race connneciton happens as follows:
1. SurfaceEdge creates new SettableSurface.
2. Downstream node is linked, adding a callback to close the current SettableSurface when downstream node is closed.
3. SurfaceEdge disconnected by UseCase lifecycle, creating a new SettableSurface.
4. Downstream node is closed by lifecycle, closing the new SettaableSureface instead of the old one.

The fix is:

1. bundling the SurfaceOutput with the SettableSurface, so they will never be out of sync, and
2) we use closure to pass the right SettableSurface to the callback.

Bug: 310040774
Bug: 324483095
Test: manual test and ./gradlew bOS
Change-Id: I3b9a54e6a2d74bf6f890b0fcf46b7d7206980c64