Add Markdown doc explaining a11y implementation details
I wrote down some of my knowledge about Compose a11y in this doc,
especially focused on details that aren't self-explanatory and that are
mostly documented in scattered patch descriptions and bugs currently.

As it's essentially a lengthy code comment explaining how to interpret
the a11y-related code, I thought it would make sense for it to live next
to those files in the codebase.  Note that .md file formatting is
supported by AndroidX's codesearch tool: see for example
https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:compose/README.md

Test: N/A
Change-Id: Ieca99e029534697aa46cd64f84c022c96d11f99b