CfW: Introduce webCommonW3C source set to fix metadata compilation while avoiding code duplication (#1123)
**The goals of this PR:**
- Enable back `compileJsWasmMainKotlinMetadata` tasks 
- Fix "common w3c" symbol resolving in IDE

**Solution:**
- The code with "common w3c" usages is moved to a new folder called
"webQuasiCommon"
- jsMain and wasmJsMain source sets include this folder

**How to test:**

These tasks should be successfull:
```
./gradlew :mpp:publishComposeJbToMavenLocal -Pcompose.platforms=web,jvm,android // no native to test on linux agents 
./gradlew :compose:mpp:demo:wasmJsBrowserProductionWebpack
./gradlew :compose:mpp:demo:jsBrowserProductionWebpack
```

**Cons:**
IDEA treats the new webQuasiCommon as a k/js source set. Meaning we
still need to be careful and test the changes (on CI too).

**Pros:***
Working with the code in IDE is more convenient now with resolved
symbols.
No need for a workaround anymore in user projects.