Fix main thread check on multiplatform
On multiplatform, `Dispatchers.Main` configurable and might be unavailable.
Currently, there is no unified API to check if current thread matches the one that's used in `Dispatchers.Main`, so using thread locals in combination with sync launch is the most reliable way to check it. If `Dispatchers.Main` is unavailable - just don't perform the check at all

Test: N/A
Change-Id: Ie07a8e23702b15d2c6967f7e1c9ad645c459195e