Providers API Surface adds BiometricPromptData
Biometric prompt data that can be optionally used by you to start the single tap sign on
flow. If you opt to use this, you can modify the behaviour of the biometric confirmation to
have occurred before the selected [CreateEntry] or [CredentialEntry] is relayed to you.

This change also introduces the AuthenticatorTypes annotation IntDef. We
reached consensus on utilizing the BiometricTeam's Framework IntDef
across teams, but on actually attempting to do so, it is not possible
since while the main type is public, the IntDef isn't.

While this change has gone extensive review, this particular change will
be accessible only to restricted access levels, and later will be
propagated to API change level.

Bug: 325469910
Test: Chained JUnit Tests
Relnote: "Adds BiometricPromptData to let providers indicate if they
want to let CredentialManager handle the biometric prompt"

Change-Id: I021fe8af199e74c140326f3a221588beab131e34