Remove unnecessary final modifier from enum methods
Metalava removes `final` modifier from methods in `final` classes to
normalize the model, and ensure consistency as historically the
underlying Java/Kotlin parsing code was inconsistent in how it
handled implicit or inherited modifiers. Unfortunately, due to a bug
in Metalava that normalization was not applied to methods in an `enum`
class (which is itself implicitly `final`). That bug has been fixed
and this is updating the few places where that affected the AndroidX
signature files.

Relnote: N/A
Bug: 326310840
Test: ./gradlew
Change-Id: I3c6d421697ad6230d9722c74b2992763701ea0c8