Refresh .ignore files after fixing Item.toString() methods
Change https://r.android.com/2973652 fixed some consistency and
correctness issues with the `Item.toString()` implementations but as
that is used for the element id in baseline files it caused some build
failures in AndroidX because previously baselined issues could not be
found (as the element ids had changed). As a result the change included
a tweak to the `Baseline` code to ignore those differences until the
baseline files could be updated.

This change updates the baseline files.

Bug: 325454211
Test: ./gradlew ignoreApiChanges; ./gradlew updateApi
Change-Id: Ic87052401d7c3c5cd9be02befdf60ea6e32eb14f