Invalidate Text layout cache even if the node is detached
Ensures that layout cache in Text modifier node is up to date even if the node is detached. Movable content can currently update the node while it is being moved, making layout cache stale as result.

Currently the order of update/attach is not a part of enforced Modifier.Node contract, and we want to make sure that such updates are supported correctly.

Bug: 327814644
Test: TextAnnotatedStringNodeTest + TextStringSimpleNodeTest
Change-Id: Icd918a71408e42c557543f33179d13ed5deab62e