Add quirk to disable abortCapturesOnStop on select platforms
The extensions implementations on certain platforms suffered from an
issue where calling abortCaptures on stop could cause
SessionProcessor.deInitSession() to hang indefinitely.

By default, CameraPipe sets abortCapturesOnStop to true on SDK levels
greater than R in order to speed up the camera switching process.
However, because of the aforementioned issue,  we need to disable this
behavior on certain platforms when it's a SessionProcessor (extension)
session.

Bug: 325088903
Bug: 327517884
Test: Test extensions on affected devices
Change-Id: I8a05c976a3097ddfc63024acac6bcda0f886e42b