Try to fix test in g3
There's NoClassDefFoundError on
androidx/window/extensions/embedding/ActivityStack$Token in g3 presubmit
when updating prebuilt library.
A possible root cause is that the library calls to ActivityStack
that with the Token property with WM Extensions library less than 5.

This CL uses the same mechanism as SplitInfo to initialize
ActivityStack. The error didn't happen on SplitInfo after all.

Bug: 326440474
Test: ./gradlew window:window:cAt on ToT and U devices
Change-Id: I2fd0e55a0d550db1bac9cf5614cee76d63802c84