Add DisplayCutoutCompat.getCutoutPath
This is mostly a straightforward exposing of the platform method in the
Compat APIs.

One subtletly is that since DisplayCutout.Builder was only added in API
33, the corresponding DisplayCutoutCompat constructor that takes a
cutout path will only add the cutout path on API 33 and above, as on API 31 and
API 32 there isn't a public way to create a platform DisplayCutout for
testing with a cutout path, even though on those APIs a "real"
DisplayCutout can include the cutout path.

Relnote: "Added DisplayCutoutCompat.getCutoutPath"

Fixes: 279635904
Test: new test

Change-Id: I58778bb5c23837304fb5130644f9f4412884cdc2