Avoid redundant posting to main thread in SystemForegroundService
`startForeground`, `cancelNotification` & `notify` are called
from the onStartCommand and onExecuted methods in SystemForegroundDispatcher
that are called on main thread anyway, so we don't need additional jumps

Additionally SystemForegroundDispatcher#handleNotify was rewritten in such way
that startForeground() is always explicitly called at the end. It is done because
startForeground() must be called for every matching parent startForegroundService()
call. After this rewrite it is easier to reason about `startForeground` call.

bug: 254846063
Test: integration app works
Change-Id: I1cdab9bd7d2c70c1acb862db8489f30c0a1ee6fc